+++ modules-services/prov-service-validation/src/main/java/org/openprovenance/prov/service/validation/ValidationServiceUtils.java
@@ -170,15 +170,25 @@ public class ValidationServiceUtils  extends ServiceUtils {
 			// save this
 			index.put(vr.getVisibleId(), vr);
 
+
+			//store report.json
+
+			report.getNamespace().register("val", "http://foo.foo/");
+			System.out.println("report.getNamespace()=" + report.getNamespace());
+			reportStorage.serializeObjectToStore(report,jsonReportStorageId);
+
+
+			// for XML Serialisation, remove all namespaces
+			report.setNamespace(null);
+			report.getValidationReport().forEach(r -> r.setNamespace(null));
+
+
 			//store report.xml
 			ByteArrayOutputStream baos=new ByteArrayOutputStream();
 			utils.getConfig().serialiser.serialiseObject(baos,report,"ignore", false);
 			nonDocumentResourceStorage.copyStringToStore(baos.toString(),reportStorageId);
 
-			//store report.json
-			ns.register("val", "http://foo.foo/");
-			report.setNamespace(ns);  //NOTE: not saved in xml version for now
-			reportStorage.serializeObjectToStore(report,jsonReportStorageId);
+
 
 			//store matrix.txt
 			String matrix = constraints.getMatrix().displayMatrix2();